Added KubernetesBroker, a wrapper around IKubernetes to allow mocking and unit testing it.
