+++ src/HookTrigger.Worker/Brokers/IKubernetesBroker.cs
@@ -1,0 +1,14 @@
+﻿using k8s.Models;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace HookTrigger.Worker.Brokers
+{
+    public interface IKubernetesBroker
+    {
+        Task<List<V1Deployment>> FindDeploymentByImageAsync(string imageName);
+
+        Task<V1Deployment> PatchNamespacedDeploymentAsync(V1Patch patch, string deploymentName, string namespaceProperty, CancellationToken cancellationToken = default);
+    }
+}
