+++ src/HookTrigger.Worker/Services/IKubernetesService.cs
@@ -5,6 +5,6 @@ namespace HookTrigger.Worker.Services
 {
     public interface IKubernetesService
     {
-        public Task PatchAllDeploymentAsync(string repoName, string tag, CancellationToken cancellationToken = default);
+        Task<int> PatchAllDeploymentAsync(string repoName, string tag, CancellationToken cancellationToken = default);
     }
 }
